apps/freelance-radar/app/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
httpx==0.26.0
pydantic==2.5.3
